final pr comments to move from interface to default abstract class
