--- smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/DefaultHttpProtocolCustomizable.kt
@@ -1,7 +1,0 @@
-package software.amazon.smithy.swift.codegen.integration
-
-abstract class DefaultHttpProtocolCustomizable : HttpProtocolCustomizable {
-    override fun baseMiddlewares(ctx: ProtocolGenerator.GenerationContext): List<OperationMiddlewareRenderable> {
-        return emptyList()
-    }
-}
