+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/HttpProtocolCustomizable.kt
@@ -7,12 +7,7 @@ import software.amazon.smithy.protocoltests.traits.HttpRequestTestCase
 import software.amazon.smithy.swift.codegen.SwiftWriter
 
 interface HttpProtocolCustomizable {
-    fun renderMiddlewares(ctx: ProtocolGenerator.GenerationContext, writer: SwiftWriter, op: OperationShape, operationStackName: String) {
-        val middlewares = operationMiddlewares(ctx)
-        for (middleware in middlewares) {
-            middleware.render(ctx, writer, ctx.service, op, operationStackName)
-        }
-    }
+    fun renderMiddlewares(ctx: ProtocolGenerator.GenerationContext, writer: SwiftWriter, op: OperationShape, operationStackName: String)
 
     fun renderInternals(ctx: ProtocolGenerator.GenerationContext) {
         // Default implementation is no-op
@@ -49,12 +44,7 @@ interface HttpProtocolCustomizable {
      * to be rendered (i.e. after integrations have had a chance to intercept). The default set of middleware for
      * a protocol can be overridden by [baseMiddlewares].
      */
-    fun operationMiddlewares(ctx: ProtocolGenerator.GenerationContext): List<OperationMiddlewareRenderable> {
-        val defaultMiddleware = baseMiddlewares(ctx)
-        return ctx.integrations.fold(defaultMiddleware) { middleware, integration ->
-            integration.customizeMiddleware(ctx, middleware)
-        }
-    }
+    fun operationMiddlewares(ctx: ProtocolGenerator.GenerationContext): List<OperationMiddlewareRenderable>
 
     fun baseMiddlewares(ctx: ProtocolGenerator.GenerationContext): List<OperationMiddlewareRenderable>
 }
