+++ smithy-swift-codegen/src/test/kotlin/MockHttpRestXMLProtocolGenerator.kt
@@ -5,7 +5,7 @@ import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.traits.TimestampFormatTrait
 import software.amazon.smithy.swift.codegen.integration.CodingKeysGenerator
 import software.amazon.smithy.swift.codegen.integration.DefaultCodingKeysGenerator
-import software.amazon.smithy.swift.codegen.integration.DefaultHttpProtocolCustomizable
+import software.amazon.smithy.swift.codegen.integration.DefaultHttpProtocolCustomizations
 import software.amazon.smithy.swift.codegen.integration.HttpBindingProtocolGenerator
 import software.amazon.smithy.swift.codegen.integration.HttpProtocolTestGenerator
 import software.amazon.smithy.swift.codegen.integration.HttpProtocolUnitTestErrorGenerator
@@ -16,14 +16,14 @@ import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
 import software.amazon.smithy.swift.codegen.integration.httpResponse.HttpResponseGeneratable
 import software.amazon.smithy.swift.codegen.integration.httpResponse.HttpResponseGenerator
 
-class MockRestXMLHttpProtocolCustomizable() : DefaultHttpProtocolCustomizable()
+class MockRestXMLHttpProtocolCustomizations() : DefaultHttpProtocolCustomizations()
 
 class MockHttpRestXMLProtocolGenerator : HttpBindingProtocolGenerator() {
     override val defaultContentType: String = "application/xml"
     override val defaultTimestampFormat: TimestampFormatTrait.Format = TimestampFormatTrait.Format.DATE_TIME
     override val protocol: ShapeId = RestXmlTrait.ID
     override val httpProtocolClientGeneratorFactory = TestHttpProtocolClientGeneratorFactory()
-    override val httpProtocolCustomizable = MockRestXMLHttpProtocolCustomizable()
+    override val httpProtocolCustomizable = MockRestXMLHttpProtocolCustomizations()
     override val codingKeysGenerator: CodingKeysGenerator = DefaultCodingKeysGenerator()
     override val httpResponseGenerator: HttpResponseGeneratable = HttpResponseGenerator(
         serviceErrorProtocolSymbol,
